2lab/requirements.txt
fastapi
uvicorn
sqlalchemy
email-validator
pydantic[email]
passlib[bcrypt]
python-jose
python-dotenv
alembic
celery
redis
websockets
aiofiles